PortoSeguro/requirements.txt
selenium
openpyxl
webdriver-manager
python-dotenv
